style: usecase 파일 분리 및 변경
